% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charliework-package.R
\docType{package}
\name{charliework-package}
\alias{charliework}
\alias{charliework-package}
\title{charliework: Sequence N for Homework 4}
\description{
Takes as input a vector x containing the first three numeric elements of this sequence and a positive (>0) integer n denoting the final nth element of the sequence to calculate.
}
\author{
\strong{Maintainer}: Charlie Bollinger \email{cb8884a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
